Simplified Collision
In order to eliminate confusion, the collision system has been simplified. Now, there is no longer a Collided and Collider. There are only Collidables. Collidables can collide with other Collidables, but cannot collide with themselves.

This implementation allows both objects to handle the collision independently (as per the GRASP principle), rather than the responsibility of the handling the collision for both objects inappropriately being assigned to only the collided.

It also enables triangular collisions, where three objects can all correctly collide with each other.

As is standard practice, the CollisionDelegate testing has been updated to accommodate for this new implementation.